Container with Most Water (C++)
When given an integer array height of length n, there are n vertical lines drawn such that the two endpoints of the i-th line are (i, 0) and (i, height[i]). So we find two lines that together with the x-axis form a container, such that the container contains the most water. It returns the maximum amount of water a container can store.