Number of People Aware of Secrets (C++)
Given an integer "delay", which means that each person will share the secret with a new person every day, starting from delay days after discovering the secret, then an integer "forget" , which means that each person will forget the secret forget days after discovering it. A person cannot share the secret on the same day they forgot it, or on any day afterwards. This returns the number of people who know the secret at the end of day n.