Best Time to Buy & Sell (C++)
You are given an array prices where prices[i] is the price of a given stock on the i-th day.

You want to maximize your profit by choosing a single day to buy one stock and choosing a different day in the future to sell that stock. This program returns the maximum profit you can achieve from this transaction. If you cannot achieve any profit, return 0.